quickstart change

+++ deep_learn_lib/losses/loss.py
@@ -22,7 +22,7 @@ class MSE(Loss):
     """
 
     def loss(self, predicted: Tensor, actual: Tensor) -> float:
-        return np.mean((predicted - actual) ** 2)
+        return float(np.mean((predicted - actual) ** 2))
 
     def grad(self, predicted: Tensor, actual: Tensor) -> Tensor:
         return 2 * (predicted - actual) / np.size(predicted)  # Added normalization
